Created "simple_oo" subpackage, new models.Query
Renamed models.Query to OOQuery
Introduced new and very different models.Query for flexible query/request handling with layer-2 Client
Introduced clients.ClientQuery and used it in clients.Client to support ResultsFromRequest objects with Client
Introduced simple_oo subpackage
Moved clients.Icinga2 and models.OOQuery (former Query) to new simple_oo.client module
Also moved base_objects and objects modules to simple_oo subpackage
Updated __init__ package imports and other imports
Took the opportunity to rewrite OOQuery, split up code into some different methods
This could coincidentally fix some bugs
The **kwargs when calling a OOQuery object do no longer support setting the name or setting init parameters for the returned objects
Updated docs
Fixed two typos
