+++ icinga2api_py/__init__.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 
 from .api import API
-from .clients import Client, StreamClient, Icinga2
+from .clients import Client, StreamClient
+from .simple_oo import Icinga2
 from .results import ResultSet, ResultsFromResponse, ResultsFromRequest, ResultList, Result
 
-__version__ = "0.6.19"
+__version__ = "0.6.20"
