+++ icinga2api_py/simple_oo/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from .client import Icinga2
